Fix all TypeScript errors and add messages.test.ts
Phase 1 Complete (130+ → 0 errors):
- Fixed exactOptionalPropertyTypes conditional spreads (~55 errors)
- Fixed seyfert type imports (6 errors)
- Fixed API call paths for audit-logs and invites (4 errors)
- Fixed wrong argument counts (6 errors)
- Fixed register.ts method name mismatches (10 errors)
- Added missing service methods (15 methods across 4 classes)
- Fixed one-off type errors (7 errors)
- Fixed audit-logs.test.ts mock structure

Phase 2 - Test Coverage Started:
- Created src/discord/messages.test.ts (12 tests)
  - sendMessage dry-run and happy path
  - editMessage, deleteMessage with confirm/dry-run gates
  - purgeMessages with confirm/dry-run gates
  - pinMessage, unpinMessage dry-run gates
  - addReaction, clearReactions dry-run gates
  - Guild allowlist enforcement

All 89 tests passing, zero TypeScript errors.

+++ src/mcp/tools/helpers.ts
@@ -81,7 +81,7 @@ export function toJsonSafe(obj: unknown): unknown {
   for (const [key, value] of entries) {
     const type = typeof value;
     if (type === 'bigint') {
-      result[key] = value.toString();
+      result[key] = (value as bigint).toString();
       changed = true;
     } else if (value !== null && type === 'object') {
       const processed = toJsonSafe(value);
